+++ app/src/main/java/io/github/daeun1012/withhotels/data/local/Hotel.kt
@@ -8,7 +8,7 @@ import com.google.gson.annotations.SerializedName
 
 @Entity(tableName = "hotels")
 data class Hotel(
-    @PrimaryKey @ColumnInfo(name = "id") val id: Int,
+    @PrimaryKey @ColumnInfo(name = "id") val id: Long,
     @field:SerializedName("name") val name: String,
     @field:SerializedName("thumbnail") val thumbnail: String,
     @field:SerializedName("description") @Embedded val description: Description? = null,
